Add ConnectionOverlay and PageSkeleton components; refactor PartnershipHub to use PartnershipHubUI
- Introduced ConnectionOverlay for user interaction with Gmail connection prompts.
- Added PageSkeleton for loading states in the Partnership Hub.
- Refactored PartnershipHub to utilize PartnershipHubUI for improved structure and maintainability.
- Implemented mock data utilities for testing and development purposes.
- Enhanced PartnershipHubSkeleton with dynamic filtering and grouping of partnerships.
